fix: twitter followers/following api
