+++ Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 description = "Reverse Engineered Twitter API"
 repository = "https://github.com/Amovane/ReverseEngineeredTwitterAPI"
 license = "MIT"
-repository = "https://github.com/Amovane/ReverseEngineeredTwitterAPI/"
+# repository = "https://github.com/Amovane/ReverseEngineeredTwitterAPI/"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -14,6 +14,6 @@ async-trait = "0.1.74"
 chrono = "0.4.26"
 dotenv = "0.15.0"
 reqwest = { version = "0.11", features = ["json", "cookies"] }
-serde = { version="1.0.164", features=["derive"] }
+serde = { version = "1.0.164", features = ["derive"] }
 serde_json = "1.0.100"
 tokio = { version = "1", features = ["full"] }
