+++ src/lib.rs
@@ -1,7 +1,8 @@
 pub mod auth;
+pub mod relation;
 pub mod search;
 pub mod types;
-pub mod relation;
+pub mod types_resp;
 use reqwest::Client;
 
 pub const LOGIN_URL: &str = "https://api.twitter.com/1.1/onboarding/task.json";
@@ -57,13 +58,13 @@ mod tests {
         let is_logged_in = api.is_logged_in().await;
         assert!(is_logged_in);
 
-        let result = search(&mut api).await;
-        assert!(result.is_ok());
+        // let result = search(&mut api).await;
+        // assert!(result.is_ok());
 
-        let res = search_tweets(&mut api).await;
-        assert!(res.is_ok());
+        // let res = search_tweets(&mut api).await;
+        // assert!(res.is_ok());
 
-        let (tweets, _) = res.unwrap();
-        assert!(tweets.len() == 0);
+        // let (tweets, _) = res.unwrap();
+        // assert!(tweets.len() == 0);
     }
 }
