+++ src/types_resp/mod.rs
@@ -1,0 +1,2 @@
+pub mod followers_types;
+pub mod following_types;
